// web/src/generated/openapi.ts
export interface components {
    schemas: {
// ...
        CreateInvitationRequest: {
            email: string;
            role: string;
        };
        CreateInvitationResponse: {
            invitation: components["schemas"]["InvitationResponse"];
            inviteUrl: string;
        };
// ...
        InvitationResponse: {
            /** Format: uuid */
            id: string;
            email: string;
            role: string;
            /** Format: date-time */
            invitedAt: string;
            /** Format: date-time */
            expiresAt: string;
            status: string;
            /** Format: uuid */
            invitedByUserId: string;
            /** Format: date-time */
            acceptedAt: null | string;
            /** Format: date-time */
            revokedAt: null | string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
// ...
    ListInvitations: {
        parameters: {
            query?: {
                sortBy?: string;
                sortOrder?: string;
                cursor?: string;
                limit?: string;
                status?: string;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["CursorPageOfInvitationResponse"];
                };
            };
            /** @description Bad Request */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/problem+json": components["schemas"]["ProblemDetails"];
                };
            };
        };
    };
// ...
}